a refactor started

+++ adv_train.py
@@ -1,12 +1,10 @@
-from typing import Any, Callable, List
-
+from typing import Any, Callable, List, Optional, cast
 import argparse
 import numpy as np
 import shutil
 import glob
 import time
 import random
-
 from tensorboardX import SummaryWriter
 
 from perceptual_advex import evaluation
@@ -16,7 +14,7 @@ from perceptual_advex.attacks import *
 from perceptual_advex.models import FeatureModel
 
 VAL_ITERS = 100
-my_device = torch.device('cuda:0') if torch.cuda.is_available() else torch.device('mps')
+
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
@@ -37,7 +35,7 @@ if __name__ == '__main__':
                         help='model to use for LPIPS distance')
     parser.add_argument('--only_attack_correct', action='store_true',
                         default=False, help='only attack examples that '
-                                            'are classified correctly')
+                        'are classified correctly')
     parser.add_argument('--randomize_attack', action='store_true',
                         default=False,
                         help='randomly choose an attack at each step')
@@ -56,7 +54,7 @@ if __name__ == '__main__':
                         help='learning rate')
     parser.add_argument('--lr_schedule', type=str, required=False,
                         help='comma-separated list of epochs when learning '
-                             'rate should drop')
+                        'rate should drop')
     parser.add_argument('--clip_grad', type=float, default=1.0,
                         help='clip gradients to this value')
 
@@ -81,8 +79,8 @@ if __name__ == '__main__':
             if args.num_epochs is None:
                 args.num_epochs = 100
         elif (
-                args.dataset.startswith('imagenet')
-                or args.dataset == 'bird_or_bicycle'
+            args.dataset.startswith('imagenet')
+            or args.dataset == 'bird_or_bicycle'
         ):
             if args.lr is None:
                 args.lr = 1e-1
@@ -98,12 +96,14 @@ if __name__ == '__main__':
     dataset, model = get_dataset_model(args)
     if isinstance(model, FeatureModel):
         model.allow_train()
-    model.to(my_device)
+
+    model.to(StaticVars.DEVICE)
 
     if args.lpips_model is not None:
         _, lpips_model = get_dataset_model(
             args, checkpoint_fname=args.lpips_model)
-        lpips_model.to(my_device)
+
+        lpips_model.to(StaticVars.DEVICE)
 
     train_loader, val_loader = dataset.make_loaders(
         workers=4, batch_size=args.batch_size)
@@ -154,18 +154,20 @@ if __name__ == '__main__':
     iteration = 0
     log_dir = os.path.join(args.log_dir, experiment_path)
     if os.path.exists(log_dir):
-        shutil.rmtree(log_dir)
-        # sleep necessary to prevent weird bug where directory isn't
-        # actually deleted
-        time.sleep(5)
+        print(f'The log directory {log_dir} exists, delete? (y/N) ', end='')
+        if not vars(args)['continue'] and input().strip() == 'y':
+            shutil.rmtree(log_dir)
+            # sleep necessary to prevent weird bug where directory isn't
+            # actually deleted
+            time.sleep(5)
     writer = SummaryWriter(log_dir)
 
     # optimizer
     optimizer: optim.Optimizer
     if args.optim == 'sgd':
         weight_decay = 1e-4 if (
-                args.dataset.startswith('imagenet')
-                or args.dataset == 'bird_or_bicycle'
+            args.dataset.startswith('imagenet')
+            or args.dataset == 'bird_or_bicycle'
         ) else 2e-4
         optimizer = optim.SGD(model.parameters(),
                               lr=args.lr,
@@ -179,7 +181,6 @@ if __name__ == '__main__':
     lr_drop_epochs = [int(epoch_str) for epoch_str in
                       args.lr_schedule.split(',')]
 
-
     # check for checkpoints
     def get_checkpoint_fnames():
         for checkpoint_fname in glob.glob(os.path.join(glob.escape(log_dir),
@@ -187,7 +188,6 @@ if __name__ == '__main__':
             epoch = int(os.path.basename(checkpoint_fname).split('.')[0])
             if epoch < args.num_epochs:
                 yield epoch, checkpoint_fname
-
 
     start_epoch = 0
     latest_checkpoint_epoch = -1
@@ -212,21 +212,21 @@ if __name__ == '__main__':
             adaptive_eps = state.get('adaptive_eps', {})
 
     # parallelize
-    device_ids = list(range(args.parallel))
-    model = nn.DataParallel(model, device_ids)
-    attacks = [nn.DataParallel(attack, device_ids) for attack in attacks]
-    validation_attacks = [nn.DataParallel(attack, device_ids)
-                          for attack in validation_attacks]
-
+    if torch.cuda.is_available():
+        device_ids = list(range(args.parallel))
+        model = nn.DataParallel(model, device_ids)
+        attacks = [nn.DataParallel(attack, device_ids) for attack in attacks]
+        validation_attacks = [nn.DataParallel(attack, device_ids)
+                              for attack in validation_attacks]
 
     # necessary to put training loop in a function because otherwise we get
     # huge memory leaks
     def run_iter(
-            inputs: torch.Tensor,
-            labels: torch.Tensor,
-            iteration: int,
-            train: bool = True,
-            log_fn: Optional[Callable[[str, Any], Any]] = None,
+        inputs: torch.Tensor,
+        labels: torch.Tensor,
+        iteration: int,
+        train: bool = True,
+        log_fn: Optional[Callable[[str, Any], Any]] = None,
     ):
         prefix = 'train' if train else 'val'
         if log_fn is None:
@@ -234,9 +234,8 @@ if __name__ == '__main__':
                 f'{prefix}/{tag}', value, iteration)
 
         model.eval()  # set model to eval to generate adversarial examples
-
-        inputs = inputs.to(my_device)
-        labels = labels.to(my_device)
+        inputs = inputs.to(StaticVars.DEVICE)
+        labels = labels.to(StaticVars.DEVICE)
 
         if args.only_attack_correct:
             with torch.no_grad():
@@ -286,9 +285,9 @@ if __name__ == '__main__':
                 else:
                     attack_name = attack.__class__.__name__
                 attack_logits = logits[
-                                attack_index * inputs.size()[0]:
-                                (attack_index + 1) * inputs.size()[0]
-                                ]
+                    attack_index * inputs.size()[0]:
+                    (attack_index + 1) * inputs.size()[0]
+                ]
                 log_fn(f'loss/{attack_name}',
                        F.cross_entropy(attack_logits, labels).item())
                 log_fn(f'accuracy/{attack_name}',
@@ -307,7 +306,6 @@ if __name__ == '__main__':
             # clip gradients and optimize
             nn.utils.clip_grad_value_(model.parameters(), args.clip_grad)
             optimizer.step()
-
 
     for epoch in range(start_epoch, args.num_epochs):
         lr = args.lr
